KHWE-78: Delete dispatch function, rename theme.js to themes.js
