+++ src/store/actionCreators/themes.js
@@ -1,12 +1,6 @@
 import { THEME_CHANGE_PALETTE_TYPE } from "../actionTypes";
 
-export function onChangeTheme(type) {
-  return dispatch => {
-    dispatch(changeTheme(type));
-  };
-}
-
-export function changeTheme(payload) {
+export function onChangeTheme(payload) {
   return {
     type: THEME_CHANGE_PALETTE_TYPE,
     payload
